backend/requirements/local.txt
# ============================================
# POSX Backend - Local Development Requirements
# ============================================

# Include production requirements
-r production.txt

# Development tools
ipython==8.18.1
ipdb==0.13.13

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Code quality
black==23.12.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1

# Django extensions
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Documentation
Sphinx==7.2.6
